interview-python-gemini/requirements.txt
# ── Web Framework ─────────────────────────────────────
fastapi==0.110.0
uvicorn==0.27.1

# ── Resume Parsing ─────────────────────────────────────
pdfplumber==0.10.3
python-docx==1.1.0

# ── Speech to Text (Free - runs locally on your PC) ───
openai-whisper==20231117

# ── LLM — Google Gemini (FREE, no credit card needed) ─
google-generativeai==0.5.2

# ── Speech Analysis ────────────────────────────────────
librosa==0.10.1
soundfile==0.12.1
nltk==3.8.1

# ── Utilities ──────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.6.1
numpy==1.26.4
